+++ frontend_tests/node_tests/lib/events.js
@@ -517,6 +517,8 @@ exports.fixtures = {
 
     restart: {
         type: "restart",
+        zulip_version: "4.0-dev+git",
+        zulip_feature_level: 55,
         server_generation: 2,
         immediate: true,
     },