+++ zerver/lib/event_schema.py
@@ -1142,7 +1142,13 @@ def check_realm_user_update(
 
 
 restart_event = event_dict_type(
-    required_keys=[("type", Equals("restart")), ("server_generation", int), ("immediate", bool)]
+    required_keys=[
+        ("type", Equals("restart")),
+        ("zulip_version", str),
+        ("zulip_feature_level", int),
+        ("server_generation", int),
+        ("immediate", bool),
+    ]
 )
 check_restart_event = make_checker(restart_event)